arch/risc-v: Refine syscall interface
