+++ arch/risc-v/src/bl602/Make.defs
@@ -23,7 +23,7 @@
 HEAD_ASRC = bl602_vectors.S
 
 # Specify our general Assembly files
-CHIP_ASRCS = bl602_head.S riscv_syscall.S bl602_entry.S
+CHIP_ASRCS = bl602_head.S bl602_entry.S
 
 CMN_ASRCS += riscv_testset.S
