+++ arch/risc-v/src/esp32c3/Make.defs
@@ -25,7 +25,7 @@ include chip/Bootloader.mk
 
 HEAD_ASRC  = esp32c3_head.S
 
-CHIP_ASRCS  = esp32c3_vectors.S esp32c3_interrupt.S riscv_syscall.S
+CHIP_ASRCS  = esp32c3_vectors.S esp32c3_interrupt.S
 
 # Specify C code within the common directory to be included
