+++ arch/risc-v/src/fe310/Make.defs
@@ -23,7 +23,7 @@
 HEAD_ASRC = fe310_vectors.S
 
 # Specify our general Assembly files
-CHIP_ASRCS = fe310_head.S riscv_syscall.S
+CHIP_ASRCS = fe310_head.S
 
 CMN_ASRCS += riscv_testset.S
