+++ arch/risc-v/src/qemu-rv32/Make.defs
@@ -23,7 +23,7 @@
 HEAD_ASRC = qemu_rv32_head.S
 
 # Specify our general Assembly files
-CHIP_ASRCS = qemu_rv32_vectors.S riscv_syscall.S
+CHIP_ASRCS = qemu_rv32_vectors.S
 
 CMN_ASRCS += riscv_testset.S
